+++ security/__init__.py
@@ -1,11 +1,13 @@
 """Modulos de autodefesa e seguranca interna do JARVIS."""
 
 from security.security_knowledge_core import SecurityKnowledgeCore
+from security.remediation_engine import RemediationEngine
 from security.security_twin import SecurityTwin
 from security.security_validation_engine import SecurityValidationEngine
 from security.threat_model_engine import ThreatModelEngine
 
 __all__ = [
+    "RemediationEngine",
     "SecurityKnowledgeCore",
     "ThreatModelEngine",
     "SecurityTwin",